Add a backtrace function, and call it on panic
Just before shutting off QEMU, print backtrace addresses on panic.

These can be mapped using `addr2line` into actual return addresses.

This backtrace function relies on the stack pointer being generated, so
note that in the function documentation and force Rust to generate
frame pointers.

`backtrace` is not marked as an unsafe function, because it relies on
constants present in codegen.